don't add vector columns for sqlite

+++ database/migrations/2026_02_09_075356_add_tag_vectors_to_tracks.php
@@ -1,6 +1,8 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Query\Grammars\PostgresGrammar;
+use Illuminate\Database\Query\Grammars\SQLiteGrammar;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
@@ -12,6 +14,11 @@ return new class extends Migration
      */
     public function up(): void
     {
+        $grammar = DB::connection()->getQueryGrammar();
+
+        // Don't apply this migration for SQLite
+        if ($grammar instanceof SQLiteGrammar) return;
+
         // set up pgvector, we need it for the new column.
         //
         // note that this may fail if the database user is not a superuser
@@ -20,14 +27,30 @@ return new class extends Migration
         // line yourself as the superuser (example for Linux):
         //
         // $ sudo -u postgres psql -c "CREATE EXTENSION IF NOT EXISTS vector;"
-        DB::statement('CREATE EXTENSION IF NOT EXISTS vector');
+        if ($grammar instanceof PostgresGrammar)
+            DB::statement('CREATE EXTENSION IF NOT EXISTS vector');
 
         // add the actual column.
-        Schema::table('tracks', function (Blueprint $table) {
-            // Laravel don't know how to emit DDL for the halfvec type.
-            DB::statement('ALTER TABLE tracks
-                ADD COLUMN tags halfvec(128) NOT NULL
-                DEFAULT \'[
+        Schema::table('tracks', function (Blueprint $table) use ($grammar) {
+            if ($grammar instanceof PostgresGrammar) {
+                // use half precision vectors. Laravel can't (as of time)
+                // emit DDL for such usage, so this has to be written as
+                // raw SQL.
+                DB::statement('ALTER TABLE tracks
+                    ADD COLUMN tags halfvec(128) NOT NULL
+                    DEFAULT \'[
+                        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
+                    ]\'');
+            } else {
+                // For other databases just use what Laravel has to offer.
+                $table->vector('tags', 128)->default('[
                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
@@ -36,7 +59,8 @@ return new class extends Migration
                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
-                ]\'');
+                ]');
+            }
         });
     }
 
@@ -45,6 +69,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        $grammar = DB::connection()->getQueryGrammar();
+
+        if ($grammar instanceof SQLiteGrammar) return;
+
         // normal drop column operation.
         Schema::table('tracks', function (Blueprint $table) {
             $table->dropColumn('tags');
@@ -53,6 +81,7 @@ return new class extends Migration
         // since we set up pgvector when migrating up, we should tear it
         // down here. again it may fail if the database user is not a
         // superuser; see above and you should know what to do.
-        DB::statement('DROP EXTENSION vector');
+        if ($grammar instanceof PostgresGrammar)
+            DB::statement('DROP EXTENSION vector');
     }
 };
